requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
sqlmodel>=0.0.8
canvasapi>=3.0.0
requests>=2.31.0
pymupdf>=1.23.0
anthropic>=0.18.0
python-multipart>=0.0.6
requests>=2.31.0
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
